To scratch yourself in the head -Ouroboros
And therefore, I was the ram herder legend mountain plateau
Now I am above the gods, below the devils, and everything in between
In the future, I will be placed in bottle caps and platinum relics by the wizards who ride upg. dragons

If intelligent aliens find the golden record, they will wonder over why it doesn‘t hold the information contained within this craft, and deem the human civilization to undeveloped as a species to be worthy of a visit, or even a place in a zoological garden. The planets will align again, and it will be overtime for the platinum records to glide after the Voyager probes.